requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
faiss-cpu>=1.9.0
streamlit>=1.29.0
PyPDF2>=3.0.1
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
tiktoken>=0.5.2
openai>=1.6.1
transformers>=4.36.2
torch>=2.1.0
numpy>=1.24.0
pandas>=2.0.0